requirements.txt
python-telegram-bot>=21.7
pytesseract>=0.3.10
Pillow>=11.0.0
spacy>=3.7.2
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.26.0SpeechRecognition>=3.10.0
pydub>=0.25.1